refactor: split bot/install modules and refresh docs
