+++ src/bot/chat-state.ts
@@ -1,0 +1,68 @@
+import type { PiSessionContext } from "../pi-session.js";
+import { getPiSessionContextKey } from "../pi-session.js";
+
+export interface BotChatState {
+  isLocallyBusy(target: PiSessionContext): boolean;
+  beginProcessing(target: PiSessionContext, promptText: string): void;
+  endProcessing(target: PiSessionContext): void;
+  beginSwitching(target: PiSessionContext): void;
+  endSwitching(target: PiSessionContext): void;
+  beginTranscribing(target: PiSessionContext): void;
+  endTranscribing(target: PiSessionContext): void;
+  getLastPrompt(target: PiSessionContext): string | undefined;
+  clearPromptMemory(target: PiSessionContext): void;
+}
+
+export function createBotChatState(): BotChatState {
+  const processingContexts = new Set<string>();
+  const switchingContexts = new Set<string>();
+  const transcribingContexts = new Set<string>();
+  const lastPrompts = new Map<string, string>();
+
+  const getContextKey = (target: PiSessionContext): string => getPiSessionContextKey(target);
+
+  return {
+    isLocallyBusy(target) {
+      const contextKey = getContextKey(target);
+      return (
+        processingContexts.has(contextKey) ||
+        switchingContexts.has(contextKey) ||
+        transcribingContexts.has(contextKey)
+      );
+    },
+
+    beginProcessing(target, promptText) {
+      const contextKey = getContextKey(target);
+      processingContexts.add(contextKey);
+      lastPrompts.set(contextKey, promptText);
+    },
+
+    endProcessing(target) {
+      processingContexts.delete(getContextKey(target));
+    },
+
+    beginSwitching(target) {
+      switchingContexts.add(getContextKey(target));
+    },
+
+    endSwitching(target) {
+      switchingContexts.delete(getContextKey(target));
+    },
+
+    beginTranscribing(target) {
+      transcribingContexts.add(getContextKey(target));
+    },
+
+    endTranscribing(target) {
+      transcribingContexts.delete(getContextKey(target));
+    },
+
+    getLastPrompt(target) {
+      return lastPrompts.get(getContextKey(target));
+    },
+
+    clearPromptMemory(target) {
+      lastPrompts.delete(getContextKey(target));
+    },
+  };
+}
