+++ test/bot/message-rendering.test.ts
@@ -1,0 +1,150 @@
+import * as formatModule from "../../src/format.js";
+
+import {
+  appendWithCap,
+  buildStreamingPreview,
+  findPreferredSplitIndex,
+  formatMarkdownMessage,
+  formatToolSummaryLine,
+  renderExtensionError,
+  renderExtensionNotice,
+  renderFailedText,
+  getWorkspaceShortName,
+  renderHelpHTML,
+  renderHelpPlain,
+  renderPromptFailure,
+  renderSessionInfoHTML,
+  renderSessionInfoPlain,
+  renderToolEndMessage,
+  renderToolStartMessage,
+  renderVoiceSupportHTML,
+  renderVoiceSupportPlain,
+  splitMarkdownForTelegram,
+  splitTelegramText,
+  stripHtml,
+  summarizeToolOutput,
+  trimLine,
+  isMessageNotModifiedError,
+  isTelegramParseError,
+} from "../../src/bot/message-rendering.js";
+
+describe("bot message rendering helpers", () => {
+  const info = {
+    sessionId: "session-1234",
+    sessionFile: "/tmp/session.jsonl",
+    workspace: "/workspace/project",
+    sessionName: "My Session",
+    modelFallbackMessage: "Using fallback model",
+    model: "anthropic/claude-sonnet-4-5",
+  };
+
+  afterEach(() => {
+    vi.restoreAllMocks();
+  });
+
+  it("renders session info and help in plain text and HTML", () => {
+    expect(renderSessionInfoPlain(info)).toContain("Session ID: session-1234");
+    expect(renderSessionInfoPlain(info)).toContain("Model note: Using fallback model");
+    expect(renderSessionInfoHTML(info)).toContain("<b>Session ID:</b>");
+    expect(renderSessionInfoHTML(info)).toContain("<code>/tmp/session.jsonl</code>");
+
+    expect(renderHelpPlain(info)).toContain("/commands — browse TelePi and Pi commands");
+    expect(renderHelpPlain(info)).toContain("Each Telegram chat/topic has its own Pi session");
+    expect(renderHelpHTML(info)).toContain("<code>/sessions &lt;path|id&gt;</code>");
+    expect(renderHelpHTML(info)).toContain("<b>Notes</b>");
+  });
+
+  it("renders voice support, tool updates, and tool summaries", () => {
+    expect(renderVoiceSupportPlain(["openai", "parakeet"]))
+      .toBe("Voice transcription: openai, parakeet.");
+    expect(renderVoiceSupportHTML([], "Missing ffmpeg")).toContain("⚠️ Missing ffmpeg");
+
+    expect(renderToolStartMessage("bash")).toEqual({
+      text: "<b>🔧 Running:</b> <code>bash</code>",
+      fallbackText: "🔧 Running: bash",
+      parseMode: "HTML",
+    });
+
+    const toolEnd = renderToolEndMessage("bash", "done", false);
+    expect(toolEnd.text).toContain("✅");
+    expect(toolEnd.text).toContain("<pre>done</pre>");
+    expect(toolEnd.fallbackText).toContain("done");
+
+    expect(formatToolSummaryLine(new Map([[
+      "read", 1,
+    ], ["bash", 2]]))).toBe("🔧 3 tools used: bash ×2, read");
+    expect(formatToolSummaryLine(new Map())).toBe("");
+  });
+
+  it("renders prompt and extension failures consistently", () => {
+    expect(renderPromptFailure("partial output", new Error("something failed")))
+      .toBe("partial output\n\n⚠️ something failed");
+    expect(renderPromptFailure("", new Error("Aborted by user"))).toBe("⏹ Aborted");
+
+    expect(renderFailedText(new Error("boom"))).toEqual({
+      text: "<b>Failed:</b> boom",
+      fallbackText: "Failed: boom",
+      parseMode: "HTML",
+    });
+
+    expect(renderExtensionNotice("Heads up", "warning")).toEqual({
+      text: "<b>⚠️</b> Heads up",
+      fallbackText: "⚠️ Heads up",
+      parseMode: "HTML",
+    });
+
+    expect(renderExtensionError("command:review", "command", "No diff found")).toEqual({
+      text: "<b>❌ /review failed:</b> No diff found",
+      fallbackText: "❌ /review failed: No diff found",
+      parseMode: "HTML",
+    });
+  });
+
+  it("splits Telegram text and markdown into safe chunks", () => {
+    const chunks = splitTelegramText(`${"a".repeat(3900)}\n${"b".repeat(3900)}`);
+    expect(chunks).toHaveLength(2);
+    expect(chunks.every((chunk) => chunk.length <= 4000)).toBe(true);
+
+    const markdown = "<".repeat(2500);
+    const renderedChunks = splitMarkdownForTelegram(markdown);
+    expect(renderedChunks.length).toBeGreaterThan(1);
+    expect(renderedChunks.every((chunk) => chunk.text.length <= 4000)).toBe(true);
+    expect(renderedChunks.map((chunk) => chunk.sourceText).join("")).toBe(markdown);
+  });
+
+  it("falls back to plain text when Telegram HTML formatting fails", () => {
+    vi.spyOn(formatModule, "formatTelegramHTML").mockImplementation(() => {
+      throw new Error("broken formatter");
+    });
+
+    expect(formatMarkdownMessage("hello <world>")).toEqual({
+      text: "hello <world>",
+      fallbackText: "hello <world>",
+      parseMode: undefined,
+    });
+  });
+
+  it("provides utility helpers for previews and string cleanup", () => {
+    expect(findPreferredSplitIndex("line1\nline2", 6)).toBe(5);
+    expect(findPreferredSplitIndex("word1 word2", 8)).toBe(5);
+    expect(findPreferredSplitIndex("abcdef", 3)).toBe(3);
+
+    expect(buildStreamingPreview("a".repeat(3801))).toContain("… streaming (preview truncated)");
+    expect(appendWithCap("abc", "def", 4)).toBe("cdef");
+    expect(summarizeToolOutput(`  ${"x".repeat(510)}  `)).toBe(`${"x".repeat(500)}\n…`);
+    expect(trimLine("one   two\nthree", 7)).toBe("one tw…");
+    expect(stripHtml("<b>Hello</b> <code>world</code>")).toBe("Hello world");
+    expect(getWorkspaceShortName("/workspace/project")).toBe("project");
+    expect(getWorkspaceShortName("C:\\workspace\\project")).toBe("project");
+  });
+
+  it("recognizes Telegram parse and message-not-modified errors", () => {
+    expect(isMessageNotModifiedError(new Error("Bad Request: message is not modified"))).toBe(true);
+    expect(isMessageNotModifiedError(new Error("other"))).toBe(false);
+
+    expect(isTelegramParseError(new Error("Bad Request: can't parse entities"))).toBe(true);
+    expect(isTelegramParseError(new Error("unsupported start tag at byte offset 1"))).toBe(true);
+    expect(isTelegramParseError(new Error("Entity name expected"))).toBe(true);
+    expect(isTelegramParseError(new Error("plain failure"))).toBe(false);
+  });
+});
